backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.8.2
python-multipart==0.0.9
httpx==0.27.0
celery==5.4.0
redis==5.0.8
python-jose[cryptography]==3.3.0
pyjwt[crypto]==2.9.0
requests==2.32.3
pandas==2.2.2
openpyxl==3.1.5
python-dateutil==2.9.0.post0
loguru==0.7.2
structlog==24.2.0
python-dotenv==1.0.1
supabase==2.6.0
starlette_exporter==0.22.0
gunicorn==22.0.0
